+++ release-notes.md
@@ -2,8 +2,9 @@
 
 ### New Features
 
-- **Discord Interagtion** — Added support for Discord Rich Presence,updating the last track information and playback
-  status.
+- **Play This Session** — Add a button to display the Play This Session window, that shows the songs that have been
+  played in the current session, and allow to go to the artist, album or song page of any of those songs or replay the
+  DJ commentary for that song.
 
 ### Bug Fixes